requirements_conda.txt
absl-py==2.4.0
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
codetiming==1.4.0
colorama==0.4.6
colorful==0.5.8
cryptography==46.0.5
datasets==4.4.2
dill==0.4.0
distlib==0.4.0
exceptiongroup==1.3.1
filelock==3.24.2
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-auth==2.48.0
googleapis-common-protos==1.72.0
grpclib==0.4.9
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
hydra-core==1.3.2
hyperframe==6.1.0
identify==2.6.16
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kimina-client==0.2.1
latex2sympy2==1.5.4
latex2sympy2_extended==1.10.2
loguru==0.7.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
math-verify==0.8.0
mdurl==0.1.2
modal==1.3.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
networkx==3.4.2
nodeenv==1.10.0
numpy==2.1.0
omegaconf==2.3.0
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
packaging>=24.0
pandas==2.3.3
peft==0.18.0
pillow==12.1.1
platformdirs==4.9.2
pre_commit==4.5.1
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.1.3
py-spy==0.4.1
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybind11==3.0.1
pycparser==3.0
Pygments==2.19.2
pylatexenc==2.10
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyvers==0.1.0
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.53.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
sentry-sdk==2.53.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sympy==1.14.0
synchronicity==0.11.1
tabulate==0.9.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensordict==0.10.0
toml==0.10.2
torchdata==0.11.0
tqdm==4.67.3
typer==0.24.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
virtualenv==20.36.1
wandb==0.23.1
watchfiles==1.1.1
Werkzeug==3.1.5
word2number==1.1
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0

